Add files via upload
Post-development PowerShell projects that I have done for companies. This PowerShell project was for KPFF Consulting Engineers company. I was working on creating a script that enhances the  installation process for AutoCAD and Civil 3D configurations, in order to reduce the amount of time is required to get this engineering programs installed.