Add files via upload
I have built this script to compare the hash value of files to ensure data integrity. I have coded this script to grab the file hash of each file, but the best part is that once it is executed it will continue going through the files and the end user can walk away from their computer. This is script is extremely useful, especially when a user is in a situation where they have to leave for the day, but needs to comb through several gigabytes of files. The script will run overnight, and by the time the end user arrives the next morning, the comparison task is completed.