Add files via upload
This script searches for the attribute "MsNPAllowDialin" for each user account within Active Directory.  I created this script because we had a user who could not log into their computer with their PIV card because the attribute "MsNPAllowDialin" was set to false. One of my colleagues, Arthuro Saalas switched the attribute to true, then the user was able to sign into their computer.  Arturo has asked me if there was a PowerShell script that would show all the user accounts that were set to false for the msNPAllowDialin attribute.